Release/gob 442

+++ src/app/store/conversation-store/effects.ts
@@ -15,6 +15,7 @@ import {
   RouterAction,
   RouterSelector
 } from '../index';
+import {InitTypeEnum} from "../../models/utils/init-type.enum";
 
 @Injectable()
 export class ConversationStoreEffects {
@@ -85,15 +86,20 @@ export class ConversationStoreEffects {
     mergeMap(() => this.messageService.leaveChat()
       .pipe(
         map((message) => {
-          if (this.firstPageElection.button) {
-            this.store.dispatch(RouterAction.buttonLogin());
-          } else if (this.firstPageElection.login) {
-            this.store.dispatch(RouterAction.loginOpen());
-          }
-          this.store.dispatch(LoginAction.leaveLogin());
+          // if (this.firstPageElection.button) {
+          //   this.store.dispatch(RouterAction.buttonLogin());
+          // } else if (this.firstPageElection.login) {
+          //   this.store.dispatch(RouterAction.loginOpen());
+          // }
+          this.store.dispatch(RouterAction.finish());
+          this.store.pipe(select(ConfigSelector.selectConfig)).subscribe(resp => {
+            if(resp.initType !== InitTypeEnum.START_CHAT_WITH_WELCOME){
+              this.store.dispatch(LoginAction.leaveLogin());
+              localStorage.removeItem('state');
+              localStorage.clear();
+            }
+          });
           this.store.dispatch(featureActions.cleanConversation());
-          localStorage.removeItem('state');
-          localStorage.clear();
           this.store.dispatch(InitWebChatAction.triggerInit({payload: true}));
           this.store.pipe(select(ConfigSelector.selectConfig), filter(fill => ((fill.preserveHistory !== undefined || fill.preserveHistory !== null)) && fill.preserveHistory))
             .subscribe(resp => {
